Added/fixed support for some fields.
The new fields are:
- first postal address as gd:structuredPostalAddress (Google API 3.0)
- birthday
- homepage
- blog
- note

The structured postal address uses the following fields:
- street
- pobox
- city
- region
- postcode
- country
- formattedAddress

When changing the address in the form field at the gmail website,
all address data is stored in formattedAddress field and all other fields
are empty (the strucutre is lost). So when such a contact gets synced with KAddressbook, I store
the address data from  formattedAddress in the street field, because in the
GUI this field is large enough to show all information.

+++ utests/utils.c
@@ -92,7 +92,7 @@ int find_load_file(char *path, char **file_content)
 	return res;
 }
 
-int find_load_photo(char *path, char **file_content, int *length)
+int find_load_photo(char *path, char **file_content, size_t *length)
 {
   	int fd, res;
 	char *file_name = path;
